genre,backdropath, rating, reviewlike add(내꺼 조아요 불가)

+++ final-pjt-be/movies/views.py
@@ -21,14 +21,16 @@ def movie_review_list(request, movie_pk):
 def movie_list(request):
     movies = Movie.objects.all()
     # 원본 코드
-    # serialzer = MovieSerialzer(movies, many=True)
-    serializer = MovieSerializer(movies, many=True, context={'request': request})
+    serializer = MovieSerializer(movies, many=True)
+    # serializer = MovieSerializer(movies, many=True, context={'request': request})
     return Response(serializer.data, status=status.HTTP_200_OK)
 
 # 영화 상세 내역 조회 원본 코드
 @api_view(['GET'])
 def movie_detail(request,movie_pk):
-    movie= Movie.objects.get(pk=movie_pk)
+    # 원본
+    # movie= Movie.objects.get(pk=movie_pk)
+    movie = get_object_or_404(Movie, pk=movie_pk)
     serializer = MovieDetailSerializer(movie)
     return Response(serializer.data, status=status.HTTP_200_OK)
 
@@ -59,13 +61,16 @@ def review_detail(request, review_pk):
         return Response({'message': f'review {review_pk} is deleted.'}, status=status.HTTP_204_NO_CONTENT)
     
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def create_review(request, movie_pk):
     # movie = Movie.objects.get(pk=movie_pk)
     movie = get_object_or_404(Movie, pk=movie_pk)
     serializer = ReviewGenSerializer(data=request.data,  context={'request': request, 'movie': movie})
     if serializer.is_valid(raise_exception=True):
-        serializer.save(movie=movie)
+        serializer.save()
+        # serializer.save(movie=movie)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
+    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
 # 영화 좋아요
 @api_view(['POST'])
@@ -89,10 +94,15 @@ def review_like_users(request, review_pk):
     review = get_object_or_404(Review, pk=review_pk)
     user = request.user
 
+ # 본인이 작성한 리뷰에는 좋아요를 누를 수 없음
+    if review.user == user:
+        return Response({'error': 'You cannot like your own review.'}, status=status.HTTP_403_FORBIDDEN)
+
     # Review_likes_users 모델을 사용하여 좋아요 추가/삭제
     review_like, created = Review_likes_users.objects.get_or_create(review=review, user=user)
 
     if not created:
+        # 이미 좋아요를 누른 경우 좋아요를 취소
         review_like.delete()
         liked = False
     else:
